Blueprint doesn't support int8 type.

+++ Source/ProjectR/Private/GameMode/RaceGameMode.cpp
@@ -140,7 +140,7 @@ void ARaceGameMode::updateJetPositions()
 	currentJetPositions = jetPositions();
 }
 
-int8 ARaceGameMode::laps()
+int ARaceGameMode::laps()
 {
 	return numberOfLaps;
 }
